Honor DriverManager#register() and driver registration thread safe
`DriverManager#register()` doesn't throw exceptions with double registrations, so this driver shouldn't as well. This commit makes the registration process honor the contract of DriverManager.

Also there was a possible race condition when registering the driver, which can be solved with applying an AtomicReference instead of a plain static field.

Fixes #1